Move the AlgoliaSearcher into a seperate package
