+++ packages/algolia-search/SearchBar.js
@@ -6,9 +6,9 @@ import debounce from "lodash/debounce";
 import { __ } from "@wordpress/i18n";
 
 /* Internal dependencies */
-import colors from "../../style-guide/colors.json";
-import { YoastButton } from "../Plugin/Shared/components/YoastButton";
-import breakpoints from "../../style-guide/responsive-breakpoints.json";
+import colors from "yoast-components/style-guide/colors.json";
+import { YoastButton } from "yoast-components/composites/Plugin/Shared/components/YoastButton";
+import breakpoints from "yoast-components/style-guide/responsive-breakpoints.json";
 
 const SearchBarWrapper = styled.div`
 	form {
