+++ packages/algolia-search/config/setupTests.js
@@ -1,0 +1,1 @@
+import "jest-styled-components";
