+++ packages/algolia-search/index.js
