+++ packages/algolia-search/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "@yoast/algolia-search",
+  "version": "1.0.0",
+  "description": "A packages that allows integration with Algolia Search",
+  "main": "index.js",
+  "repository": "https://github.com/Yoast/javascript",
+  "author": "Yoast",
+  "license": "GPL",
+  "private": false,
+  "dependencies": {
+    "@wordpress/a11y": "^1.0.7",
+    "@wordpress/i18n": "^1.1.0",
+    "algoliasearch": "^3.22.3",
+    "lodash": "^4.17.4",
+    "prop-types": "^15.6.0",
+    "react": "^16.2.0",
+    "react-intl": "^2.4.0",
+    "styled-components": "^2.1.2",
+    "yoast-components": "^4.22.1"
+  },
+  "devDependencies": {
+    "@babel/preset-env": "^7.3.4",
+    "@babel/preset-react": "^7.0.0",
+    "babel-jest": "^24.5.0",
+    "babel-loader": "^7.1.1",
+    "babel-plugin-styled-components": "^1.5.1",
+    "babel-preset-react": "^6.24.1",
+    "enzyme": "^3.3.0",
+    "enzyme-to-json": "^3.3.3",
+    "jest": "^24.5.0",
+    "jest-styled-components": "^5.0.1",
+    "react-test-renderer": "^16.8.4"
+  },
+  "scripts": {
+    "test": "jest"
+  },
+  "jest": {
+    "testRegex": ".*Test.js$",
+    "transform": {
+      "^.+\\.jsx?$": "babel-jest"
+    },
+    "snapshotSerializers": [
+      "enzyme-to-json/serializer"
+    ],
+    "setupFilesAfterEnv": [
+      "<rootDir>/config/setupTests.js"
+    ]
+  }
+}
