+++ packages/algolia-search/tests/AlgoliaSearcherTest.js
@@ -1,6 +1,6 @@
 import React from "react";
-import { createComponentWithIntl } from "../../../utils/intlProvider";
-import AlgoliaSearcher from "../AlgoliaSearcher.js";
+import { createComponentWithIntl } from "yoast-components/utils/intlProvider";
+import AlgoliaSearcher from "../index.js";
 
 test( "the AlgoliaSearcher component with headingText matches the snapshot", () => {
 	const component = createComponentWithIntl(
