+++ packages/algolia-search/tests/LoadingTest.js
