+++ packages/yoast-components/package.json
@@ -73,7 +73,8 @@
     "styled-components": "^2.1.2",
     "whatwg-fetch": "^1.0.0",
     "wicked-good-xpath": "^1.3.0",
-    "yoastseo": "^1.49.1"
+    "yoastseo": "^1.49.1",
+    "@yoast/algolia-search": "^1.0.0"
   },
   "devDependencies": {
     "autoprefixer": "^6.4.0",
